chore: Update examples
