+++ examples/Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["axum", "simple", "procmacro"]
+members = ["axum", "simple"]
 resolver = "2"
 
 [workspace.package]
@@ -7,5 +7,4 @@ version = "0.1.0"
 edition = "2021"
 
 [workspace.dependencies]
-fluent-static = { path = "../fluent-static" }
-fluent-static-codegen = { path = "../fluent-static-codegen" }
+fluent-static = { path = "../crates/lib" }
