--- examples/axum/build.rs
@@ -1,9 +1,0 @@
-use fluent_static_codegen::{generate, MessageBundleCodeGenerator};
-
-pub fn main() {
-    generate!(
-        "./l10n",
-        MessageBundleCodeGenerator::new("en-US"),
-        "l10n.rs"
-    );
-}
