+++ examples/axum/src/main.rs
@@ -1,6 +1,17 @@
 use axum::{routing::get, Router};
-use fluent_static::axum::RequestLanguage;
+use fluent_static::support::axum::RequestLanguage;
 use maud::{html, Markup};
+
+mod l10n {
+    use fluent_static::message_bundle;
+    #[message_bundle(
+        resources = [
+            ("l10n/en-US/messages.ftl", "en-US"), 
+            ("l10n/fr-CH/messages.ftl", "fr-CH"), 
+        ],
+        default_language = "en-US")]
+    pub struct Messages;
+}
 
 #[tokio::main]
 async fn main() {
@@ -9,7 +20,7 @@ async fn main() {
     axum::serve(listener, app).await.unwrap();
 }
 
-async fn handler(RequestLanguage(msgs): RequestLanguage<l10n::messages::MessagesBundle>) -> Markup {
+async fn handler(RequestLanguage(msgs): RequestLanguage<l10n::Messages>) -> Markup {
     let name = "Guest";
     html! {
         html {
@@ -20,12 +31,9 @@ async fn handler(RequestLanguage(msgs): RequestLanguage<l10n::messages::Messages
             }
             body {
                 h1 {
-                    (msgs.hello(name).unwrap())
+                    (msgs.hello(name))
                 }
             }
         }
     }
-}
-mod l10n {
-    fluent_static::include_source!("l10n.rs");
 }
