--- examples/procmacro/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "procmacro"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-fluent-static = { workspace = true }
-
-[build-dependencies]
-fluent-static-codegen = { workspace = true }
