--- examples/procmacro/l10n/en-US/messages.ftl
@@ -1,4 +1,0 @@
-hello = Hello, { $name }
-
-simple = Simple
-  .attribute = Simple Attribute
