--- examples/procmacro/l10n/fr-CH/messages.ftl
@@ -1,4 +1,0 @@
-hello=Bonjour, { $name }
-
-simple = Simple
-  .attribute = Simple Attribute
