+++ examples/simple/Cargo.toml
@@ -1,10 +1,7 @@
 [package]
-name = "example"
+name = "procmacro"
 version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-fluent-static = { workspace = true }
-
-[build-dependencies]
-fluent-static-codegen = { workspace = true }
+fluent-static = { workspace = true, features = ["icu"] }
