+++ examples/simple/l10n/fr-CH/messages.ftl
@@ -2,3 +2,5 @@ hello=Bonjour, { $name }
 
 simple = Simple
   .attribute = Simple Attribute
+
+total_amount = Total: { NUMBER($amount, currencyDisplay:"symbol") }
